requirements.txt
django
djangorestframework
django-jazzmin
psycopg2
django-environ
django-health-check